+++ test/test_helper.rb
@@ -14,12 +14,15 @@ I18n.default_locale = :en
 
 $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 
+require_relative 'stub_autodev'
+
 require 'autodev/errors'
 require 'autodev/logger'
 require 'autodev/config_validator'
 require 'autodev/project_validator'
 require 'autodev/app_validator'
 require 'autodev/app_instructions'
+require 'autodev/screenshot_uploader'
 require 'autodev/config'
 require 'autodev/language_detector'
 require 'autodev/locales'